Move API in its own package, change module name to allow for importing
